fix: rename baseGenerator
