+++ lib/api/src/version.ts
@@ -1,1 +1,1 @@
-export const version = '6.0.0-beta.24';
+export const version = '6.0.0-beta.25';
