+++ lib/cli/src/generators/EMBER/index.ts
@@ -1,4 +1,4 @@
-import { baseGenerator, Generator } from '../generator';
+import { baseGenerator, Generator } from '../baseGenerator';
 
 const generator: Generator = async (packageManager, npmOptions, options) => {
   baseGenerator(packageManager, npmOptions, options, 'ember', {
